#472: Move FDO .NET from old CLR syntax to the new CLR syntax
git-svn-id: http://svn.osgeo.org/fdo/trunk@4423 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/GenericRdbms/Managed/Src/OSGeo/FDO/Providers/Rdbms/Override/MySQL/mgOvPropertyMappingSingle.cpp
@@ -37,6 +37,6 @@ NAMESPACE_OSGEO_FDO_PROVIDERS_RDBMS_OVERRIDE_MYSQL::OvPropertyMappingSingle::OvP
 
 NAMESPACE_OSGEO_FDO_PROVIDERS_RDBMS_OVERRIDE_MYSQL::OvPropertyMappingSingle::OvPropertyMappingSingle() : NAMESPACE_OSGEO_FDO_PROVIDERS_RDBMS_OVERRIDE::OvPropertyMappingSingle(System::IntPtr::Zero, false)
 {
-	EXCEPTION_HANDLER(Attach((FdoIDisposable*)FdoMySQLOvPropertyMappingSingle::Create(), true))
+	EXCEPTION_HANDLER(Attach(IntPtr((FdoIDisposable*)FdoMySQLOvPropertyMappingSingle::Create()), true))
 }
